# swagger.yaml
openapi: 3.0.0
info:
  title: ATM CleanGuard API Documentation
  version: 1.0.0
  description: ATM CleanGuard is a digital complaint management system for reporting ATM cleanliness and maintenance issues.
servers:
  - url: /v1
    description: API v1
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /atms/{atm_id}/complaints:
    get:
      summary: Get complaint history of ATM
      tags: [ATM Locations]
      parameters:
        - in: path
          name: atm_id
          required: true
          schema: { type: string }
      responses:
        "200": { description: List of anonymized complaints }
  /complaints:
    post:
      summary: Submit a new complaint
      tags: [Complaints]
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                atm_id: { type: string }
                description: { type: string }
                complaint_type: { type: string, enum: [dirty, ac_issue, garbage, damage, no_power, other] }
                photo: { type: string, format: binary }
      responses:
        "201": { description: Submission successful }
    get:
      summary: Get all my complaints
      tags: [Complaints]
      parameters:
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer }
      responses:
        "200": { description: OK }
  /complaints/{complaint_id}:
    get:
      summary: Get Complaint Details
      tags: [Complaints]
      parameters:
        - in: path
          name: complaint_id
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK }
    put:
      summary: Update complaint
      tags: [Complaints]
      parameters:
        - in: path
          name: complaint_id
          required: true
          schema: { type: string }
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                description: { type: string }
                photo: { type: string, format: binary }
      responses:
        "200": { description: OK }
    delete:
      summary: Delete a complaint
      tags: [Complaints]
      parameters:
        - in: path
          name: complaint_id
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK }
...
  /admin/complaints:
    get:
      summary: View all complaints for this bank branch
      tags: [Admin]
      parameters:
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: atm_id
          schema: { type: string }
        - in: query
          name: from_date
          schema: { type: string, format: date }
        - in: query
          name: to_date
          schema: { type: string, format: date }
        - in: query
          name: page
          schema: { type: integer }
      responses:
        "200": { description: OK }
  /admin/complaints/{complaint_id}:
    get:
      summary: Complaint detail (admin)
      tags: [Admin]
      parameters:
        - in: path
          name: complaint_id
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK }
...